Fix detection of multiple discrete GPUs in DeviceInfo
Prior to this change only a single discrete GPU would be detected and
present in `DeviceInfo`.

This occurred as in the case of multiple discrete GPUS, the
`ExternalVideo` struct would be overwritten within the loop walking
the PCI devices. Only a single append of the final state would be
performed at the end of the loop (if applicable).

The fix is to create a new `ExternalVideo` struct for each iteration
and append before performing the next iteration (if applicable).